// base/api/generated/models/DbiamCreatePersonWithPersonenkontexteBodyParams.ts
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { DbiamCreatePersonenkontextBodyParams } from './DbiamCreatePersonenkontextBodyParams';
import {
    DbiamCreatePersonenkontextBodyParamsFromJSON,
    DbiamCreatePersonenkontextBodyParamsFromJSONTyped,
    DbiamCreatePersonenkontextBodyParamsToJSON,
} from './DbiamCreatePersonenkontextBodyParams';

/**
 * 
 * @export
 * @interface DbiamCreatePersonWithPersonenkontexteBodyParams
 */
export interface DbiamCreatePersonWithPersonenkontexteBodyParams {
    /**
     * 
     * @type {string}
     * @memberof DbiamCreatePersonWithPersonenkontexteBodyParams
     */
    familienname: string;
    /**
     * 
     * @type {string}
     * @memberof DbiamCreatePersonWithPersonenkontexteBodyParams
     */
    vorname: string;
    /**
     * 
     * @type {string}
     * @memberof DbiamCreatePersonWithPersonenkontexteBodyParams
     */
    personalnummer?: string;
    /**
     * 
     * @type {Date}
     * @memberof DbiamCreatePersonWithPersonenkontexteBodyParams
     */
    befristung?: Date;
    /**
     * 
     * @type {Array<DbiamCreatePersonenkontextBodyParams>}
     * @memberof DbiamCreatePersonWithPersonenkontexteBodyParams
     */
    createPersonenkontexte: Array<DbiamCreatePersonenkontextBodyParams>;
}

/**
 * Check if a given object implements the DbiamCreatePersonWithPersonenkontexteBodyParams interface.
 */
export function instanceOfDbiamCreatePersonWithPersonenkontexteBodyParams(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "familienname" in value;
    isInstance = isInstance && "vorname" in value;
    isInstance = isInstance && "createPersonenkontexte" in value;

    return isInstance;
}

export function DbiamCreatePersonWithPersonenkontexteBodyParamsFromJSON(json: any): DbiamCreatePersonWithPersonenkontexteBodyParams {
    return DbiamCreatePersonWithPersonenkontexteBodyParamsFromJSONTyped(json, false);
}

export function DbiamCreatePersonWithPersonenkontexteBodyParamsFromJSONTyped(json: any, ignoreDiscriminator: boolean): DbiamCreatePersonWithPersonenkontexteBodyParams {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'familienname': json['familienname'],
        'vorname': json['vorname'],
        'personalnummer': !exists(json, 'personalnummer') ? undefined : json['personalnummer'],
        'befristung': !exists(json, 'befristung') ? undefined : (new Date(json['befristung'])),
        'createPersonenkontexte': ((json['createPersonenkontexte'] as Array<any>).map(DbiamCreatePersonenkontextBodyParamsFromJSON)),
    };
}

export function DbiamCreatePersonWithPersonenkontexteBodyParamsToJSON(value?: DbiamCreatePersonWithPersonenkontexteBodyParams | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'familienname': value.familienname,
        'vorname': value.vorname,
        'personalnummer': value.personalnummer,
        'befristung': value.befristung === undefined ? undefined : (value.befristung.toISOString()),
        'createPersonenkontexte': ((value.createPersonenkontexte as Array<any>).map(DbiamCreatePersonenkontextBodyParamsToJSON)),
    };
}