// base/api/generated/models/DbiamUpdatePersonenkontexteBodyParams.ts
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { DbiamPersonenkontextBodyParams } from './DbiamPersonenkontextBodyParams';
import {
    DbiamPersonenkontextBodyParamsFromJSON,
    DbiamPersonenkontextBodyParamsFromJSONTyped,
    DbiamPersonenkontextBodyParamsToJSON,
} from './DbiamPersonenkontextBodyParams';

/**
 * 
 * @export
 * @interface DbiamUpdatePersonenkontexteBodyParams
 */
export interface DbiamUpdatePersonenkontexteBodyParams {
    /**
     * Date of the most recent changed personenkontext
     * @type {Date}
     * @memberof DbiamUpdatePersonenkontexteBodyParams
     */
    lastModified?: Date | null;
    /**
     * The amount of personenkontexte
     * @type {number}
     * @memberof DbiamUpdatePersonenkontexteBodyParams
     */
    count: number;
    /**
     * 
     * @type {Array<DbiamPersonenkontextBodyParams>}
     * @memberof DbiamUpdatePersonenkontexteBodyParams
     */
    personenkontexte: Array<DbiamPersonenkontextBodyParams>;
}

/**
 * Check if a given object implements the DbiamUpdatePersonenkontexteBodyParams interface.
 */
export function instanceOfDbiamUpdatePersonenkontexteBodyParams(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "count" in value;
    isInstance = isInstance && "personenkontexte" in value;

    return isInstance;
}

export function DbiamUpdatePersonenkontexteBodyParamsFromJSON(json: any): DbiamUpdatePersonenkontexteBodyParams {
    return DbiamUpdatePersonenkontexteBodyParamsFromJSONTyped(json, false);
}

export function DbiamUpdatePersonenkontexteBodyParamsFromJSONTyped(json: any, ignoreDiscriminator: boolean): DbiamUpdatePersonenkontexteBodyParams {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'lastModified': !exists(json, 'lastModified') ? undefined : (json['lastModified'] === null ? null : new Date(json['lastModified'])),
        'count': json['count'],
        'personenkontexte': ((json['personenkontexte'] as Array<any>).map(DbiamPersonenkontextBodyParamsFromJSON)),
    };
}

export function DbiamUpdatePersonenkontexteBodyParamsToJSON(value?: DbiamUpdatePersonenkontexteBodyParams | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'lastModified': value.lastModified === undefined ? undefined : (value.lastModified === null ? null : value.lastModified.toISOString()),
        'count': value.count,
        'personenkontexte': ((value.personenkontexte as Array<any>).map(DbiamPersonenkontextBodyParamsToJSON)),
    };
}