// base/api/generated/models/LockUserBodyParams.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LockUserBodyParams
 */
export interface LockUserBodyParams {
    /**
     * 
     * @type {boolean}
     * @memberof LockUserBodyParams
     */
    lock: boolean;
    /**
     * 
     * @type {string}
     * @memberof LockUserBodyParams
     */
    lockedBy: string;
    /**
     * Required if Befristung is set
     * @type {Date}
     * @memberof LockUserBodyParams
     */
    lockedUntil?: Date;
}

/**
 * Check if a given object implements the LockUserBodyParams interface.
 */
export function instanceOfLockUserBodyParams(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "lock" in value;
    isInstance = isInstance && "lockedBy" in value;

    return isInstance;
}

export function LockUserBodyParamsFromJSON(json: any): LockUserBodyParams {
    return LockUserBodyParamsFromJSONTyped(json, false);
}

export function LockUserBodyParamsFromJSONTyped(json: any, ignoreDiscriminator: boolean): LockUserBodyParams {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'lock': json['lock'],
        'lockedBy': json['locked_by'],
        'lockedUntil': !exists(json, 'locked_until') ? undefined : (new Date(json['locked_until'])),
    };
}

export function LockUserBodyParamsToJSON(value?: LockUserBodyParams | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'lock': value.lock,
        'locked_by': value.lockedBy,
        'locked_until': value.lockedUntil === undefined ? undefined : (value.lockedUntil.toISOString()),
    };
}