// base/api/generated/models/PersonenkontexteUpdateResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { DBiamPersonenkontextResponse } from './DBiamPersonenkontextResponse';
import {
    DBiamPersonenkontextResponseFromJSON,
    DBiamPersonenkontextResponseFromJSONTyped,
    DBiamPersonenkontextResponseToJSON,
} from './DBiamPersonenkontextResponse';

/**
 * 
 * @export
 * @interface PersonenkontexteUpdateResponse
 */
export interface PersonenkontexteUpdateResponse {
    /**
     * 
     * @type {Array<DBiamPersonenkontextResponse>}
     * @memberof PersonenkontexteUpdateResponse
     */
    dBiamPersonenkontextResponses: Array<DBiamPersonenkontextResponse>;
}

/**
 * Check if a given object implements the PersonenkontexteUpdateResponse interface.
 */
export function instanceOfPersonenkontexteUpdateResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "dBiamPersonenkontextResponses" in value;

    return isInstance;
}

export function PersonenkontexteUpdateResponseFromJSON(json: any): PersonenkontexteUpdateResponse {
    return PersonenkontexteUpdateResponseFromJSONTyped(json, false);
}

export function PersonenkontexteUpdateResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonenkontexteUpdateResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'dBiamPersonenkontextResponses': ((json['dBiamPersonenkontextResponses'] as Array<any>).map(DBiamPersonenkontextResponseFromJSON)),
    };
}

export function PersonenkontexteUpdateResponseToJSON(value?: PersonenkontexteUpdateResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'dBiamPersonenkontextResponses': ((value.dBiamPersonenkontextResponses as Array<any>).map(DBiamPersonenkontextResponseToJSON)),
    };
}