// base/api/generated/models/PersonFrontendControllerFindPersons200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { PersonendatensatzResponse } from './PersonendatensatzResponse';
import {
    PersonendatensatzResponseFromJSON,
    PersonendatensatzResponseFromJSONTyped,
    PersonendatensatzResponseToJSON,
} from './PersonendatensatzResponse';

/**
 * 
 * @export
 * @interface PersonFrontendControllerFindPersons200Response
 */
export interface PersonFrontendControllerFindPersons200Response {
    /**
     * 
     * @type {number}
     * @memberof PersonFrontendControllerFindPersons200Response
     */
    total: number;
    /**
     * 
     * @type {number}
     * @memberof PersonFrontendControllerFindPersons200Response
     */
    offset: number;
    /**
     * 
     * @type {number}
     * @memberof PersonFrontendControllerFindPersons200Response
     */
    limit: number;
    /**
     * 
     * @type {Array<PersonendatensatzResponse>}
     * @memberof PersonFrontendControllerFindPersons200Response
     */
    items: Array<PersonendatensatzResponse>;
}

/**
 * Check if a given object implements the PersonFrontendControllerFindPersons200Response interface.
 */
export function instanceOfPersonFrontendControllerFindPersons200Response(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "total" in value;
    isInstance = isInstance && "offset" in value;
    isInstance = isInstance && "limit" in value;
    isInstance = isInstance && "items" in value;

    return isInstance;
}

export function PersonFrontendControllerFindPersons200ResponseFromJSON(json: any): PersonFrontendControllerFindPersons200Response {
    return PersonFrontendControllerFindPersons200ResponseFromJSONTyped(json, false);
}

export function PersonFrontendControllerFindPersons200ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonFrontendControllerFindPersons200Response {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'total': json['total'],
        'offset': json['offset'],
        'limit': json['limit'],
        'items': ((json['items'] as Array<any>).map(PersonendatensatzResponseFromJSON)),
    };
}

export function PersonFrontendControllerFindPersons200ResponseToJSON(value?: PersonFrontendControllerFindPersons200Response | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'total': value.total,
        'offset': value.offset,
        'limit': value.limit,
        'items': ((value.items as Array<any>).map(PersonendatensatzResponseToJSON)),
    };
}