// base/api/generated/models/PersonLockResponse.ts
/* eslint-disable */
/**
 * dBildungs IAM
 * The dBildungs IAM server API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PersonLockResponse
 */
export interface PersonLockResponse {
    /**
     * 
     * @type {string}
     * @memberof PersonLockResponse
     */
    message: string;
}

/**
 * Check if a given object implements the PersonLockResponse interface.
 */
export function instanceOfPersonLockResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "message" in value;

    return isInstance;
}

export function PersonLockResponseFromJSON(json: any): PersonLockResponse {
    return PersonLockResponseFromJSONTyped(json, false);
}

export function PersonLockResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonLockResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'message': json['message'],
    };
}

export function PersonLockResponseToJSON(value?: PersonLockResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'message': value.message,
    };
}